refactor: unify analyze command with clean architecture
Replace monolithic analyze.go (1,450 lines) with clean architecture
implementation that delegates to app.AnalyzeUseCase, following
repository principles of thin CLI layer with service delegation.

Changes:
- Replace old analyze.go with refactored version using app.AnalyzeUseCase
- Remove analyze_refactored.go after promoting to main implementation
- Move isInteractiveEnvironment() to utils.go to eliminate duplication
- Rename types: AnalyzeCommandRefactored → AnalyzeCommand
- Maintain full CLI compatibility (flags, behavior, output format)

Benefits:
- 73% code reduction: 1,450 → 514 lines in analyze.go
- Proper separation: CLI → UseCase → Service → Domain
- Eliminates duplicate command registration and flag handling
- Consolidates progress tracking, error categorization, parallel execution
- Builder pattern for dependency injection via AnalyzeUseCaseBuilder

All tests pass, CLI behavior unchanged, no regressions.
